+++ src/tubee.client.ts
@@ -1,7 +1,7 @@
 const { V1Api, HttpBasicAuth, localVarRequest } = require('@gyselroth/tubee-sdk-node');
 const fs = require('fs');
 const keytar = require('keytar');
-import { keytarPath, keytarPathOrig, Config, ConfigStore, configPath, tubectlFolder } from './config';
+import { keytarPath, keytarPathOrig, Config, Context, ConfigStore, configPath, tubectlFolder } from './config';
 
 if (!fs.existsSync(tubectlFolder)) {
   fs.mkdirSync(tubectlFolder);
@@ -21,7 +21,14 @@ export default class TubeeClient {
    * Factory
    */
   public async factory(category: string, options = null) {
-    const config: Config = ConfigStore.get(options);
+    var config: Context;
+
+    try {
+      config = ConfigStore.get(options);
+    } catch(Err) {
+      console.error('you need to login first (tubectl login)');
+      process.exit();
+    }
 
     if (config.debug || options.debug) {
       localVarRequest.debug = true;
@@ -32,7 +39,7 @@ export default class TubeeClient {
     }
 
     var server = config.url || 'https://localhost:8090';
-    var password = await keytar.getPassword('tubee', config.username || 'admin');
+    var password = await keytar.getPassword('tubee', config.name || 'admin');
     var client = new V1Api(server);
     var basic = new HttpBasicAuth();
     basic.username = config.username || 'admin';